refactor: relocate files across layers and clean up imports
